파싱하기(float출력 안되는거 고침)

+++ src/main/java/com/example/hello/parser/HospitalParser.java
@@ -52,7 +52,8 @@ public class HospitalParser implements Parser<Hospital>{
         hospital.setHealthcareProviderCount(Integer.parseInt(row[29]));
         hospital.setPatientRoomCount(Integer.parseInt(row[30]));
         hospital.setTotalNumberOfBeds(Integer.parseInt(row[31]));
-        hospital.setTotalAreaSize(Integer.parseInt(row[32]));
+        hospital.setTotalAreaSize( Float.parseFloat(row[32]));
+
 
         return hospital;
     }
